Move DonationForm subcomponents
This helps to keep the different donation form variants (that will grow
over time until we find a better way to break them up) separate from the
components used inside the forms.
