+++ banners/desktop/styles/styles.pcss
@@ -19,8 +19,8 @@ weight should be component < theme < banner.
 @import '../../../components/ButtonClose/ButtonClose.pcss';
 @import '../../../components/DonationForm/DonationForm.pcss';
 @import '../../../components/ProgressBar/ProgressBar.pcss';
-@import '../../../components/DonationForm/SelectGroup.pcss';
-@import '../../../components/DonationForm/SelectCustomAmount.pcss';
+@import '../../../components/DonationForm/subcomponents/SelectGroup.pcss';
+@import '../../../components/DonationForm/subcomponents/SelectCustomAmount.pcss';
 @import '../../../components/DonationForm/2-step-form.pcss';
 @import '../../../components/SoftClose/SoftClose.pcss';
