+++ banners/english/styles/styles.pcss
@@ -18,8 +18,8 @@ weight should be component < theme < banner.
 @import '../../../components/BannerTransition/BannerTransition.pcss';
 @import '../../../components/ButtonClose/ButtonClose.pcss';
 @import '../../../components/DonationForm/DonationForm.pcss';
-@import '../../../components/DonationForm/SelectGroup.pcss';
-@import '../../../components/DonationForm/SelectCustomAmount.pcss';
+@import '../../../components/DonationForm/subcomponents/SelectGroup.pcss';
+@import '../../../components/DonationForm/subcomponents/SelectCustomAmount.pcss';
 @import '../../../components/DonationForm/2-step-form.pcss';
 @import '../../../components/ProgressBar/ProgressBar.pcss';
