+++ banners/mobile/styles/styles.pcss
@@ -23,9 +23,9 @@ weight should be component < theme < banner.
 @import '../../../components/Footer/SelectionInput.pcss';
 @import '../../../components/ButtonClose/ButtonClose.pcss';
 @import '../../../components/DonationForm/DonationForm.pcss';
-@import '../../../components/DonationForm/SelectGroup.pcss';
-@import '../../../components/DonationForm/SelectCustomAmount.pcss';
-@import '../../../components/DonationForm/SmsBox.pcss';
+@import '../../../components/DonationForm/subcomponents/SelectGroup.pcss';
+@import '../../../components/DonationForm/subcomponents/SelectCustomAmount.pcss';
+@import '../../../components/DonationForm/subcomponents/SmsBox.pcss';
 @import '../../../components/DonationForm/2-step-form.pcss';
 @import '../../../components/Slider/Slider.pcss';
 @import '../../../components/ProgressBar/ProgressBar.pcss';
